[FIX] 캐시 수정 및 추가
