+++ build.gradle
@@ -97,7 +97,12 @@ dependencies {
     implementation 'io.micrometer:micrometer-registry-prometheus' // prometheus 추가
 
     // SpringBatch
-    // implementataion 'org.springframework.boot:spring-boot-starter-batch'
+    implementation 'org.springframework.boot:spring-boot-starter-batch'
+    // jsoup
+    implementation 'org.jsoup:jsoup:1.18.1'
+
+    // logstash
+    implementation 'net.logstash.logback:logstash-logback-encoder:7.0'
 
 }
