+++ src/main/java/com/sparta/quokkatravel/domain/common/cache/InvalidateAccommodationCache.java
@@ -1,4 +1,5 @@
-package com.sparta.quokkatravel.domain.common.aop;
+package com.sparta.quokkatravel.domain.common.cache;
+
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,5 +8,5 @@ import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
-public @interface InvalidateRoomCache {
+public @interface InvalidateAccommodationCache {
 }
