+++ src/main/java/com/sparta/quokkatravel/domain/common/cache/InvalidateRoomCache.java
@@ -1,5 +1,4 @@
-package com.sparta.quokkatravel.domain.common.aop;
-
+package com.sparta.quokkatravel.domain.common.cache;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -8,5 +7,5 @@ import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
-public @interface InvalidateAccommodationCache {
+public @interface InvalidateRoomCache {
 }
